% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\docType{data}
\name{Connect}
\alias{Connect}
\title{Open a session to a server running Connect}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Connect
}
\value{
A Connection Object
}
\description{
Open a session to a server running Connect
}
\keyword{datasets}
