% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{ping_connect}
\alias{ping_connect}
\title{Ping a connect server}
\usage{
ping_connect(url, port)
}
\arguments{
\item{url}{the url of connect}

\item{port}{the port of connect}
}
\value{
The status code of the ping
}
\description{
Ping a connect server
}
\examples{
\dontrun{
ping_connect("url", "3939")
}
}
