% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspm.R
\docType{data}
\name{RSPM}
\alias{RSPM}
\title{Open a session to a server running Connect}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RSPM
}
\value{
A Connection Object
}
\description{
Open a session to a server running Connect
}
\keyword{datasets}
